// vcasino-admin/src/app/models/esport/ITournamentStatistics.ts
export interface ITournamentStatistics {
   tournamentId: number;
   title: string;
   discipline: string;
   tournamentPage: string;
   image: string;
   matchCount: number;
   startDate: string;
   endDate: string;
   betCount: number;
   totalAmountWagered: number;
   totalAmountWin: number;
   totalAmountLoss: number;
   profit: number;
}